api/requirements.txt
appdirs==1.4.3
argon2-cffi==16.3.0
certifi==2017.4.17
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==2.0
django-filter==1.0.4
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
django-sslserver==0.20
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
greenlet==0.4.12
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
msgpack-python==0.4.8
neovim==0.2.0
openapi-codec==1.3.1
packaging==16.8
Pillow==5.0.0
psycopg2==2.7.1
pycparser==2.18
PyJWT==1.5.2
pyparsing==2.2.0
pytz==2017.3
requests==2.18.1
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
